requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure Services
azure-ai-formrecognizer==3.3.2
azure-ai-documentintelligence==1.0.0b1
openai==1.3.0

# OCR Libraries
paddleocr==2.7.3
easyocr==1.7.1
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.1

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers>=2.5.0
faiss-cpu==1.7.4

# LLM & RAG
langchain==0.0.340
langchain-community==0.0.1
tiktoken==0.5.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2

# Development
pytest==7.4.3
httpx==0.25.2
black==23.11.0
flake8==6.1.0
